[#49] Set up navigator and screen structure
