+++ integration_test/utils/test_util.dart
@@ -9,7 +9,7 @@ class TestUtil {
   /// localization, routes, etc)
   static Widget pumpWidgetWithRealApp(String initialRoute) {
     _initDependencies();
-    return MyApp();
+    return SurveyApp();
   }
 
   /// We normally use this function to test a specific [widget] without
